+++ Assets/Scripts/Tanks/AITank.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 15ad47177a53345b5b1d44437c17e270
+guid: aa1c00ef9023e44c69364518a38c197b
 timeCreated: 1501275918
 licenseType: Free
 MonoImporter:
