+++ Assets/Scripts/Tanks/BaseTank.cs
@@ -1,21 +1,21 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 
-public class BaseTank : MonoBehaviour
-{
-	public float health = 100f;
+[RequireComponent(typeof(TankStats))]
+public class BaseTank : MonoBehaviour {
 
-	public float GetHealth()
-	{
-		return health;
+	protected TankStats stats;
+
+	public virtual void Start() {
+		stats = GetComponent<TankStats>();
+	}
+	
+	public float GetHealth(){
+		return stats.Health;
 	}
 
-	public void TakeDamage(float amount)
-	{
-		health -= amount;
-		if (health <= 0)
-		{
+	public void TakeDamage(float amount) {
+		stats.Health -= amount;
+		if (stats.Health <= 0) {
 			GameManager.instance.RemovePlayer(this.gameObject);
 			Destroy(gameObject);
 		}
