+++ Assets/Scripts/Tanks/BaseTank.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 95379669d1aec4911b12a81f28ca0a7e
+guid: 168ced63be51d4258a3e387e3bd20858
 timeCreated: 1501334543
 licenseType: Free
 MonoImporter:
