+++ Assets/Scripts/Tanks/TankStats.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b02a55cfebf434ed48478dc5910d5b21
-timeCreated: 1501259845
+guid: fefac242275e44f2fbf0bf7355f48d16
+timeCreated: 1501409091
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
